fix: return score

+++ QuiSNCF/Repository/PlayerRepository.cs
@@ -184,7 +184,7 @@ public class PlayerRepository(GameDbContext db, ILogger<PlayerRepository> logger
         var baseScore = CalculateScore(dto.Tries);
         //var streakMultiplier = await GetStreakMultiplier(dto.Name, gameType);
         var isNewPlayer = !await HasPlayedBeforeSpecialWeek(dto.Name, gameType);
-        var finalScore = (int)(baseScore * (isNewPlayer ?(float)1.5 : 1));
+        var finalScore = (int)(baseScore * (isNewPlayer ? 1.5 : 1.0));
 
 
         var dailyPlay = new DailyPlay
